refactor: split string builtin lowering
