+++ src/codegen/lower_inst/builtins/strings/simple.rs
@@ -1,0 +1,191 @@
+//! Purpose:
+//! Lowers direct string transforms, trim variants, HTML escaping, and first-byte case changes.
+//!
+//! Called from:
+//! - The string builtin lowering facade.
+//!
+//! Key details:
+//! - Direct runtime calls share coercion helpers while specialized results retain PHP boxing.
+
+use super::*;
+
+/// Lowers a one-argument string builtin that directly delegates to a runtime helper.
+pub(crate) fn lower_unary_string_runtime(
+    ctx: &mut FunctionContext<'_>,
+    inst: &Instruction,
+    name: &str,
+    runtime_label: &str,
+) -> Result<()> {
+    load_single_string_arg(ctx, inst, name)?;
+    abi::emit_call_label(ctx.emitter, runtime_label);
+    store_if_result(ctx, inst)
+}
+
+/// Lowers `htmlspecialchars()` / `htmlentities()` — escapes the subject string (operand 0).
+/// `name` is the calling builtin's PHP name, used in argument-coercion diagnostics. The
+/// optional `flags` and `encoding` arguments are accepted (so the common `htmlspecialchars($s,
+/// ENT_QUOTES)` call form compiles) but not applied: `__rt_htmlspecialchars` implements the
+/// ENT_QUOTES behaviour, which matches PHP's default flag set and the overwhelmingly-common
+/// ENT_QUOTES call. (A flag-aware runtime — doctype-dependent `&apos;` vs `&#039;` — is a follow-up.)
+pub(crate) fn lower_html_escape(
+    ctx: &mut FunctionContext<'_>,
+    inst: &Instruction,
+    name: &str,
+) -> Result<()> {
+    let ptr_reg = string_ptr_reg(ctx);
+    let len_reg = string_len_reg(ctx);
+    load_string_arg_to_regs(ctx, inst, 0, name, ptr_reg, len_reg)?;
+    abi::emit_call_label(ctx.emitter, "__rt_htmlspecialchars");
+    store_if_result(ctx, inst)
+}
+
+/// Lowers `grapheme_strrev()` and boxes its `string|false` result as `Mixed`.
+pub(crate) fn lower_grapheme_strrev(ctx: &mut FunctionContext<'_>, inst: &Instruction) -> Result<()> {
+    load_single_string_arg(ctx, inst, "grapheme_strrev")?;
+    abi::emit_call_label(ctx.emitter, "__rt_grapheme_strrev");
+    box_grapheme_strrev_result(ctx);
+    store_if_result(ctx, inst)
+}
+
+/// Lowers `ucfirst()` by copying the string and uppercasing the first ASCII byte.
+pub(crate) fn lower_ucfirst(ctx: &mut FunctionContext<'_>, inst: &Instruction) -> Result<()> {
+    load_single_string_arg(ctx, inst, "ucfirst")?;
+    abi::emit_call_label(ctx.emitter, "__rt_strcopy");
+    emit_first_char_case_adjust(ctx, "ucfirst", 97, 122, FirstCharAdjust::Uppercase);
+    store_if_result(ctx, inst)
+}
+
+/// Lowers `lcfirst()` by copying the string and lowercasing the first ASCII byte.
+pub(crate) fn lower_lcfirst(ctx: &mut FunctionContext<'_>, inst: &Instruction) -> Result<()> {
+    load_single_string_arg(ctx, inst, "lcfirst")?;
+    abi::emit_call_label(ctx.emitter, "__rt_strcopy");
+    emit_first_char_case_adjust(ctx, "lcfirst", 65, 90, FirstCharAdjust::Lowercase);
+    store_if_result(ctx, inst)
+}
+
+/// Lowers `trim()`/`ltrim()`/`rtrim()`/`chop()` for default and explicit masks.
+pub(crate) fn lower_trim_like(
+    ctx: &mut FunctionContext<'_>,
+    inst: &Instruction,
+    name: &str,
+    default_runtime_label: &str,
+    mask_runtime_label: &str,
+) -> Result<()> {
+    if inst.operands.is_empty() || inst.operands.len() > 2 {
+        return Err(CodegenIrError::invalid_module(format!(
+            "{} expected 1 or 2 args, got {}",
+            name,
+            inst.operands.len()
+        )));
+    }
+    let ptr_reg = string_ptr_reg(ctx);
+    let len_reg = string_len_reg(ctx);
+    load_string_arg_to_regs(ctx, inst, 0, name, ptr_reg, len_reg)?;
+    if inst.operands.len() == 1 {
+        abi::emit_call_label(ctx.emitter, default_runtime_label);
+    } else {
+        lower_trim_mask_arg(ctx, inst, name)?;
+        abi::emit_call_label(ctx.emitter, mask_runtime_label);
+    }
+    store_if_result(ctx, inst)
+}
+
+/// Lowers a two-argument string builtin that directly delegates to a runtime helper.
+pub(crate) fn lower_binary_string_runtime(
+    ctx: &mut FunctionContext<'_>,
+    inst: &Instruction,
+    name: &str,
+    runtime_label: &str,
+) -> Result<()> {
+    load_binary_string_args(ctx, inst, name)?;
+    abi::emit_call_label(ctx.emitter, runtime_label);
+    store_if_result(ctx, inst)
+}
+/// Describes how the first-byte ASCII case helper mutates matched characters.
+pub(super) enum FirstCharAdjust {
+    Uppercase,
+    Lowercase,
+}
+
+/// Boxes the raw `grapheme_strrev()` runtime result as PHP `string|false`.
+pub(super) fn box_grapheme_strrev_result(ctx: &mut FunctionContext<'_>) {
+    let false_label = ctx.next_label("grapheme_strrev_false");
+    let done_label = ctx.next_label("grapheme_strrev_done");
+
+    match ctx.emitter.target.arch {
+        Arch::AArch64 => {
+            ctx.emitter.instruction(&format!("cbz x1, {}", false_label));       // box false when grapheme scanning reports a null string pointer
+            crate::codegen::emit_box_current_value_as_mixed(ctx.emitter, &PhpType::Str);
+            ctx.emitter.instruction(&format!("b {}", done_label));              // skip false boxing after a successful grapheme reversal
+            ctx.emitter.label(&false_label);
+            ctx.emitter.instruction("mov x1, #0");                              // false payload = 0 for grapheme_strrev() failure
+            ctx.emitter.instruction("mov x2, #0");                              // bool mixed payloads do not use a high word
+            ctx.emitter.instruction("mov x0, #3");                              // runtime tag 3 = bool false
+            abi::emit_call_label(ctx.emitter, "__rt_mixed_from_value");
+            ctx.emitter.label(&done_label);
+        }
+        Arch::X86_64 => {
+            ctx.emitter.instruction("test rax, rax");                           // test the returned string pointer for the failure sentinel
+            ctx.emitter.instruction(&format!("jz {}", false_label));            // box false when grapheme scanning reports a null string pointer
+            crate::codegen::emit_box_current_value_as_mixed(ctx.emitter, &PhpType::Str);
+            ctx.emitter.instruction(&format!("jmp {}", done_label));            // skip false boxing after a successful grapheme reversal
+            ctx.emitter.label(&false_label);
+            ctx.emitter.instruction("xor edi, edi");                            // false payload = 0 for grapheme_strrev() failure
+            ctx.emitter.instruction("xor esi, esi");                            // bool mixed payloads do not use a high word
+            ctx.emitter.instruction("mov eax, 3");                              // runtime tag 3 = bool false
+            abi::emit_call_label(ctx.emitter, "__rt_mixed_from_value");
+            ctx.emitter.label(&done_label);
+        }
+    }
+}
+
+/// Emits target-aware first-byte ASCII case adjustment for `ucfirst()` and `lcfirst()`.
+pub(super) fn emit_first_char_case_adjust(
+    ctx: &mut FunctionContext<'_>,
+    label_prefix: &str,
+    lower_bound: u8,
+    upper_bound: u8,
+    adjust: FirstCharAdjust,
+) {
+    match ctx.emitter.target.arch {
+        Arch::AArch64 => {
+            let done = ctx.next_label(&format!("{}_done", label_prefix));
+            ctx.emitter.instruction(&format!("cbz x2, {}", done));              // leave empty strings unchanged because there is no first byte
+            ctx.emitter.instruction("ldrb w9, [x1]");                           // load the first byte of the copied string for ASCII case checks
+            ctx.emitter.instruction(&format!("cmp w9, #{}", lower_bound));      // compare the first byte against the lower ASCII case bound
+            ctx.emitter.instruction(&format!("b.lt {}", done));                 // leave bytes below the case range unchanged
+            ctx.emitter.instruction(&format!("cmp w9, #{}", upper_bound));      // compare the first byte against the upper ASCII case bound
+            ctx.emitter.instruction(&format!("b.gt {}", done));                 // leave bytes above the case range unchanged
+            match adjust {
+                FirstCharAdjust::Uppercase => {
+                    ctx.emitter.instruction("sub w9, w9, #32");                 // convert lowercase ASCII to uppercase
+                }
+                FirstCharAdjust::Lowercase => {
+                    ctx.emitter.instruction("add w9, w9, #32");                 // convert uppercase ASCII to lowercase
+                }
+            }
+            ctx.emitter.instruction("strb w9, [x1]");                           // store the adjusted first byte into the copied string
+            ctx.emitter.label(&done);
+        }
+        Arch::X86_64 => {
+            let done = ctx.next_label(&format!("{}_done", label_prefix));
+            ctx.emitter.instruction("test rdx, rdx");                           // leave empty strings unchanged because there is no first byte
+            ctx.emitter.instruction(&format!("jz {}", done));                   // skip first-byte mutation for empty strings
+            ctx.emitter.instruction("movzx ecx, BYTE PTR [rax]");               // load the first byte of the copied string for ASCII case checks
+            ctx.emitter.instruction(&format!("cmp cl, {}", lower_bound));       // compare the first byte against the lower ASCII case bound
+            ctx.emitter.instruction(&format!("jb {}", done));                   // leave bytes below the case range unchanged
+            ctx.emitter.instruction(&format!("cmp cl, {}", upper_bound));       // compare the first byte against the upper ASCII case bound
+            ctx.emitter.instruction(&format!("ja {}", done));                   // leave bytes above the case range unchanged
+            match adjust {
+                FirstCharAdjust::Uppercase => {
+                    ctx.emitter.instruction("sub cl, 32");                      // convert lowercase ASCII to uppercase
+                }
+                FirstCharAdjust::Lowercase => {
+                    ctx.emitter.instruction("add cl, 32");                      // convert uppercase ASCII to lowercase
+                }
+            }
+            ctx.emitter.instruction("mov BYTE PTR [rax], cl");                  // store the adjusted first byte into the copied string
+            ctx.emitter.label(&done);
+        }
+    }
+}
